*) Major update to improve parallelization, especially for the first gravitational consistency stage. *) Fixed pair of rare bugs that connected unrelated halos together in the same tree; this affected 5 links per million (thanks to Kristin Riebe and Christoph Behrens for finding).  However, the maximum forest size should now be more manageable in many cases. *) New option to apply fix for Rockstar spins (affecting versions *prior* to Rockstar v0.99.9-RC3); search README for (FIX_ROCKSTAR_SPINS) usage details. *) Includes conversion code for generating forests in the Sussing Merger Trees format; see README for usage details. *) New option to limit memory usage (set LIMIT_MEMORY=1 in config file), helpful when dealing with >2048^3 boxes.  Currently, this turns off gzipping outputs (which requires forked processes to remain alive longer), and also limits the parallelization when resorting outputs (which would require more snapshots to be in memory at the same time). *) Experimental option for binary intermediate output formats, which reduces load/run times for large boxes; search README for INPUT_FORMAT for details. *) New PDF readme.

+++ src/distance.c
@@ -1,9 +1,9 @@
 #include <math.h>
 #include <assert.h>
 
-#define MAX_Z 100.0
-#define Z_BINS 1000.0
-#define TOTAL_BINS ((int)(MAX_Z*Z_BINS))
+#define MAX_Z 100
+#define Z_BINS 1000
+#define TOTAL_BINS (MAX_Z*Z_BINS)
 double Omega_M, Omega_L, h;
 double Dh;
 double _Dc[TOTAL_BINS];
@@ -33,9 +33,9 @@ void init_cosmology(double omega_m, double omega_l, double h0)
   h = h0;
   Dh = 2997.92458 / h; //In Mpc  (Speed of light / H0)
   for (i=0; i<TOTAL_BINS; i++) {
-    z = (float)(i+0.5) / Z_BINS;
+    z = (float)(i+0.5) / (double)Z_BINS;
     _Dc[i] = Dc_int * Dh;
-    Dc_int += 1.0 / (_E(z)*Z_BINS);
+    Dc_int += 1.0 / (_E(z)*(double)Z_BINS);
   }
 }
 
@@ -48,7 +48,7 @@ double scale_factor(double z) {
 }
 
 double comoving_distance(double z) {
-  float f = z*Z_BINS;
+  float f = z*(double)Z_BINS;
   int bin = f;
   if (z<0) return 0;
   if (bin>(TOTAL_BINS-2)) return (_Dc[TOTAL_BINS-1]);
